seperated implementation of hardware and onboard modules
